+++ src/Symfony/Bundle/FrameworkBundle/Command/EventDispatcherDebugCommand.php
@@ -22,7 +22,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\EventDispatcher\ListenerIntrospectionInterface;
 use Symfony\Contracts\Service\ServiceProviderInterface;
 
 /**
@@ -157,7 +157,7 @@ class EventDispatcherDebugCommand extends Command
         }
     }
 
-    private function searchForEvent(EventDispatcherInterface $dispatcher, string $needle): array
+    private function searchForEvent(ListenerIntrospectionInterface $dispatcher, string $needle): array
     {
         $output = [];
         $lcNeedle = strtolower($needle);